feat(deps): updated deps
This is to prepare for Java11 update. In summary, we will need some additional deps for XML and these libraries would have conflicts with some old versions of some libraries that we are currently using. For these reason, these libraries are updated and the additional deps that we will need are included.