fix(bunq): Add session handler
This session handler mimics what we do in the authenticator's method isSessionActive(). Once we get to that method during authentication we've already cleared the session storage which could be the reason for failing refreshes. Checking this during keepAlive will keep the session state in case successful. Worst case this doesn't help, but then a session exception is thrown which was the previous default behaviour (through SessionHandler.alwaysFail()).