fix(bunq): Throw SESSION_EXPIRED for broken credentials
We get an error when registering the device if we've registered before without using a wildcard to allow requests from all IP addresses. Credentials created before the change to use wildcard needs to be put in AUTH_ERROR so that we trigger a new SCA and re-register the device with the wildcard.