fix(PersistentStorage): Fix storage for AgentPlatform
AgentPlatform stores values in PersistentStorage as serialized strings. In PersistentStorageService we use .put(String, String) method to store values in storage. Before those changes values nested inside an object were not added to the secretValuesSubject which is used fot finding values which need to be hashed in S3 logs. In result access token and other sensitive values were store in logs. After this change similar to method .put(String, Object) all of the values will be extracted and added separately to secretValuesSubject