feat(SBAB/legacy): Make agent testable through wiremock
Since the agent is legacy we have to apply the legacy host strategy in order for the agent to not make requests against the real bank when running a wiremock test. Also refactored so none of the three clients use the client directly as the strategy then won't be applied.