chore(as-scaling): Pre-scaling for pay peak
Change request
===

[AAP-1486]

General information
---

**Summary of change:**

We cannot trust our autoscaling metrics and need to scale up to 100 minReplicas
during the week of the 7th of September.

Blast radius
---

**What is the potential negative impact on our customers (and other
services)?** Sometimes blast radius can be reduced by rolling out to a subset
of users/API calls using an `if` condition, feature toggle, Unleash, dark
launching techniques etc.

There is a correlation between the number of aggregation service pods and or
the number of EC2 instances and outages in certain other infrastructure, for
example the Secret Service and the Aggregation Controller nginx proxy.

**How will you measure the impact on production and potential failure?**

* Monitor essential systems
** Aggregation Service
** Aggregatiion Controller
** Secret Service
** Authentication Engine
** Aggregation Engine

**How will this change be rolled back if this breaks?**

* `kubectl -n aggregation edit <agggregation hpa>` and set minReplicas back to 50
* New PR reverting the change
* If a dependent service fails rolling restarts of those services have been
  successful mitigations previously

Checklist
---

 * [x] I have adhered to the [Tink code review
   best-practises](https://book-of.tink.network/engineering/software/code_review/).
 * [x] If other teams may be affected by the changes, I have assigned them as
   reviewers.
 * [x] The code I'm changing has a clear code owner. If not, please help out to
   find and [assign
   one](https://github.com/tink-ab/tink-backend/blob/master/README_CODEOWNERS.md)!
 * [ ] **I have verified that this PR works in production.** (check this after
   merging the PR)

If you are unsure/in disagreement about the solution/code, bring in an
additional reviewer!