refactor(Swedbank/OB/SE): Improve detailed consent handling
Throw an NO_ACCOUNTS login error if we don't get any accounts. No need to get detailed consent for nothing, and since the all accounts consent is only valid once we'll get an error in the account fetcher later on. Also check that the consent is actually valid before storing it.