fix(DanskeBankCreditCardFetcherTest): Fix flaky AccountIdentifiers test
At the time of writing this message, the location for flaky test failing line is
se.tink.backend.aggregation.agents.nxgen.serviceproviders.banks.danskebank.fetchers.DanskeBankCreditCardFetcherTest.shouldFetchCreditCardAccounts(DanskeBankCreditCardFetcherTest.java:122).

The problem is that in recursive comparison AssertJ does not respect overridden equals() methods.
Because of that, BbanIdentifier and NorwegianIdentifier are treated as equal, as they both have only
accountNumber field set and it's set to the same value.

[ITE-2838]