fix(DanskeBankCreditCardFetcherTest): Fix flaky AccountIdentifiers test
1. First issue - identifiers are not treated as Set

During recursive comparison, when AssertJ extracts all object's fields to compare,
firstly it tries to find a field's getter - only if getter doesn't exists it chooses a field
value itself.

In se.tink.backend.aggregation.nxgen.core.account.Account we have a Set of
AccountIdentifiers and there is a getter that returns them as a list. Depending on some
random stuff, this list might have identifiers in different order, which causes recursive
comparison to fail.

The solution is to rename this method so it's no longer treated as getter - we have to
break "get" + <fieldName in camelCase> pattern. Then, identifiers would be treated as
Set and their order won't matter anymore.

2. Second issue - duplicated identifiers require custom comparator

In se.tink.backend.aggregation.nxgen.core.account.Account identifiers are duplicated
in 2 fields - "identifiers" and "idModule.identifiers". Now, let's say they both were assigned
a random order {A,B,C,D} and identifiers build in unit test got {A,B,D,C} order.

AssertJ compares actual & expected by using double for loop - it checks every actual value
with all expected values to see if anything matches. Whenever it find actual-expected pair,
the expected value is removed in order to not match it twice.

During first comparison (for "idModule.identifiers" field), it will compare:
A-A
B-B
C-D,C-C
D-D
and all those comparisons would be marked as "visited". For some weird reason, a visited
comparison is omitted and effectively treated as "equal".

So, in second comparison (for "identifiers" field) it will go on with:
A-A
B-B
C-D -> visited, so equal
D-C -> not visited and not equal
and the test fails.

A solution to this problem is to assert identifiers list separately, e.g. using custom comparator.

[ITE-2838]