refactor(Skandia/RE): Add common headers and verify in wiremock test
ADRUM and ADRUM_1 header is set on all requests that go to https://client-apis.skandia.se but we omitted it for some requests for unknown reasons. To keep it consistent and close to how the app works it's been added to the requests and also added to aap file to verify that the headers are properly set.