fix(Skandia/RE): Handle expired refresh token
With new host and endpoints Skandia seems to return {"httpCode":"401","httpMessage":"Unauthorized","moreInformation":"Cannot pass the security checks that are required by the target API or operation, Enable debug headers for more details."} when the refresh token has expired. Since this affects some users now this is a quick fix that just looks at the status code and does not touch existing logic.