refactor(event): check decision strategy on interceptor level
Previously we were checking whether we should emit an event or not in event producer component and returning empty Optional if decision-strategy decides to not produce event. The problem with this approach is, we still try to read the response body input stream to deserialize the stream into a String in the interceptor, which is redundant and also introduces an additional risk as it changes the natural flow of how HttpClient works during an agent operation.