refactor(auth-abort): fixed a bad enum naming
ABORTING_OPERATION_FAILED sounds as if we tried to abort and failed, but it reality we set this status at the end of an operation when it ends without abortion so it does not necessarily mean that we even tried to abort. For this reason, the naming was misleading and I fixed that.