feat(agentWiremockTest): command to load session from credentials
`loadLoginSession()` must be invoked for the agent to load the persistent & session storage from the Credentials. This is done as part of Authentication::isLoggedIn(), but must be done explicitly like this in case the agent test skip authentication.

This command is to be used when tests are built to skip authentication.