fix(agentWiremockTest): Load agent session if auth is skipped
The wiremock test (builder) allows agent tests to run only authentication, only fetching/refresh/pis or both.

In the case a test runs fetching/refresh/pis only (without authentication) the persistent & session storages and http client cookies (if any) are not loaded from the Credentials::sensitivePayload.

This means that if a wiremock test builder is setup with e.g. .addPersistentStorageData("foo", "bar") the agent would not see this data.

The fix for this is to load the agent's session from the Credentials::sensitivePayload by invoking loadLoginSession for tests that skip authentication.