feat(fi-spankki-ob): Decrease BGR rate limit to avoid 429
This might not help at all but let's keep the reasoning persisted.

Spankki has 2 limits:
- 900 req/min
- 60 req/s
If we exceed them, we should try again after 1 minute or 1 second accordingly.

Assumptions:
- 1 http exchange with API takes 100ms+
- that gives max 10 req/s and 60 req/min per agent
- so we can have max 6 agents/s and 15 agents/min
- 1 full refresh takes 20 seconds on average (this might not be accurate)
- during 20 seconds, we cannot have more than 6 agents at any second
- that gives 6/20 = 0.3 BGR/s => 0.003 BGR/s per pod (we have 100 pods)

Current BGR is 0.025 so let's just decrease it 10 times to 0.0025 (<0.003) and see what happens.
BGR already takes 30min so now it can take up to 5hours but it shouldn't be a big problem.

[ITE-2742]