fix(handelsbanken/re): fix for parsing kapital investment
[TC-5488] - after switching to new version (3.6.11.4), need to support
proper parsing of KF investments (there is no ambassador, that why
 we figured it out after deployment from s3 logs). custodyAccountNumber
was used as unique id and it was deleted from json with new version.
accountNumberFormatted is added with new version and can be used as
unique id (and it is the same as custodyAccountNumber), just need
to delete spaces.