fix(Lansforsakringar): disabled test
Disabled Lansforsakringar Wiremock test because the runtime dependency is pointing a wrong path. To fix that, first a granular Bazel package should be created for the agent and then this package should be included as a runtime dependency for the test. This work is outside of the scope of this PR so I am ignoring the test instead.