test(Fiducia): ignored Wiremock tests
After changing dependency chain to ensure that each Wiremock test will include only its agent as a runtime dependency (not all agents) we started to miss a runtime dependency. By adding "//src/integration/lib/src/main/java/se/tink/backend/aggregation/agents/agentfactory" as a dependency (hence including all agents as a dependency as it was the case before) solves the issue but this is not a proper solution so instead of doing that I ignored the test. If market team decides that this workaround is better than ignoring the test they might go this way.