fix(WORKSPACE/deps): upgrade commons-lang3 dep
There are two defined deps for `org.apache.commons:commons-lang3` in WORKSPACE. One is version `3.9` and the other `3.4`.

This commit upgrades the second to version `3.9` as well.

Previous changes ran into a dependency shadowing issue as some deps were linked/compiled in a different order. The issue was that a method in this library was used and it was introduced in version `3.5`, which caused runtime/test issues (MethodNotFound).

The `maven_install.json` file was updated running `$ bazel run @unpinned_maven//:pin`