fix(se/crosskey): fixed duplicate issue
We are apparently retrying to fetch transactions if we are going outside of the allowed interval. That is a problem if the user is not present because we can only get transactions for 90 days.

In that case, we get one page correctly -> then we ask for a page outside the interval → we get an error saying we’re outside the interval → we retry with the dates we had in the first page → we get duplicates.

I removed the retry and check if the user is present or not to know when we should only ask for 90 days of transactions.

TC-5513