chore(autoscaling): increase req/s metrics twice
Because of not proper autoscaling (requests towards AS instead of actual usage) - we scales our AS in the night
when we actually not need it for only short period of time.
Because of that most of bgr refreshes - are processed on 75 pod and currently it looks to be enough.

As a short term solution - to decrease the cost rates - we can increase from 1.5 - 3 req/s.
And think of some better solution. (maybe scaling only on CPU usage is enough) - that needs to be verified.

[AAP-1750]