feat(agent_sdk): generic steppable execution
Previously(/currently) the authentication is designed around "steps" (small units of code to execute sequentially).

We do however want to use steps for other parts of the sdk, e.g. Payments. This new package is a generic implementation of steppable execution that will allow us to reuse it for auth and payments.

It consists of (roughly):

- BaseStep that all other steps extends - this allows us to treat all steps the same and execute them using the same code
- InteractiveStep that can do everything (interact with the user, declare that the flow is done)
- IntermediateStep that can interact with the user and go to other steps, it CANNOT signal that the flow is done
- NonInteractiveStep is a subset of InteractiveStep, it CANNOT interact with the user
- ExecutionFlow that is used to compose steps into a working program

The step implementations (Interactive, Intermediate and NonInteractive) have a strict API that requires users of said classes to adhear to the type signature, e.g. a step cannot declare that it's done without returning data of a certain type.
The ExecutionFlow also has this strict API which prohibits users from composing steps of different types.

IntermediateSteps are meant to be used by framework steps that can be used for any "flow return type" (e.g. ConsentStatus) as they never, and cannot, return done.

Note: in order to achieve a strict API I had to (or I couldn't find another solution to the problem) split the steps up into Interactive and Intermediate steps.
Note2: The steps extend BaseStep, and the different type of Response/Request classes, using the `final` keyword. That means that no agent implementation can modify the API contract by overriding e.g. `executeInternal`.