fix(ICS): Set userIp to empty string if user is not present
From ICS docs: "The PSU’s IP address if the PSU is currently logged in with the TPP. It would be blank in case the PSU is not current logged in.". The header is optional, but keeping this logic as it was before the use of Operation/User (since it worked).