fix(se/swedbank): added profile check
added logic to make sure chosen bank also have to have a privateprofile , otherwise we will get error from bank when trying to reach endpoints that we are not authorized to. Only in case of fallback & where customer has both banks. Since fallback makes it OK for user to see both savingsbank and swedbank accounts and profiles, something that is not possible in RE or OB or Business.

Used constant bankName in profileParameters instead of raw string - sonar issue

TC-5667