chore(amex): add offset for over 1000 tranactions
We have a problem when we had assume that we could get more transactions during a time span by incresing the limit to over 1000 transactions.
This is not possible do to AMEX having a upper limit on 1000 transactions. I have now added so we will first ask for 1000 transaction and if we
get that during this time the user have more we will use an offset of 1000 with a limit of 1000 to get the rest