fix(fts/payment): propagate debtor info
[NZG-1132]

We weren't transfering it before, probably in error, and now it is a hard requirement to get acc-check part working.
Furthermore, I fixed us clearing the info on payment object after create - this is probably not a big deal, as this info doesn't get propagated to rest of tink at that time, but still better to keep it clean.

Payment class is a mess.