feat(agent_sdk): runtime steppable execution loop
Execute one or more steps with a given StepRequest. The execution loop will always return if the response indicates that it requires user interaction or it's done.

Future improvements can be to cap the number of executions in the loop and/or for how long time it has executed. This way we can mitigate long running agents, e.g. in a bank polling scenario.