refactor(se/skandia-ob): added key pagination for agent
The agent had implemented date paginator but infact the actual implementation lacked pagination and the agent was only making one call to the transactions endpoint to fetch 3 months of transaction. In the AIS v2 release of the APIs from the bank, they implemented the key pagination. The agent was updated to use the v2 version of APIs in #13970 . This commit adds the support of key pagination in the agent.
[TC-2671]