test(agent_sdk/payment): always return PENDING on first getStatus
Changed the test agent to always return signing results in the state PENDING the first time the method getStatus is invoked, next time it's invoked the sign result will be the predetermined one (from the ctor argument).

This way we can be sure that the tests always run at least 2 polling attempts and that the final status of a payment cannot be PENDING unless the polling timeout is hit.