feat(wiremock): Add bulk payment support to payment commands
If a bulk payment has been set the PaymentCommand will try to execute a bulk payment, otherwise check that payment has been set and proceed with the code that was already there.
Not adding any bulk payment support in the PaymentGBCommand for now since it checks specifically that the agents implements TypedPaymentControllerable (just added a payment null check).